known words stat
